fix errors

+++ webrtc-sys/src/audio_track.cpp
@@ -156,8 +156,8 @@ void AudioTrackSource::set_audio_options(
 }
 
 void AudioTrackSource::on_captured_frame(rust::Slice<const int16_t> audio_data,
-                                         int sample_rate,
-                                         size_t number_of_channels,
+                                         uint32_t sample_rate,
+                                         uint32_t number_of_channels,
                                          size_t number_of_frames) const {
   source_->on_captured_frame(audio_data, sample_rate, number_of_channels,
                              number_of_frames);
